Correctly calculate week number as per ISO week definition
First day of the week does not always belong to that week. However,
Thursday is guaranteed to be on the correct week number. So let's make
sure that for the purpose of figuring out the week number, we always
check Thursday of that week. If confused, re-read the Wikipedia page and
apply beverage of your choice.

Fixes #8
Fixes #4 (properly this time)