Introduce yearly notes page
This change adds a page for writing yearly notes.
There's one year notes page per year the calendar spans.
I've also adjusted how the header of the year overview page is rendered
to better support calendars that are not full years (Jan - Dec), but
span, for example, a semester instead. In which case, it'd automatically
say something like "September 2026 - July 2027".

Closes https://github.com/klimeryk/recalendar.js/pull/24

Signed-off-by: Jakov Petrina <jpetrina@nyqcd.org>