Add support for all languages that dayjs supports
This partially fixes issue like #96. It will "only" translate the days
of the week and month names. But with other customizations, you can get
pretty close to a fully translated generated calendar.
The UI for the configuration will be still in English, though.

Still, makes sense to unlock this feature, cause it will allow folks to
generate calendars in any language they want, including ones that are
usually very neglected or marginalized.