Switch to AGPL license
It's "compatible" with GPLv3 I was using previously, but it has
additional clauses that make it especially suited for a project like
recalendar. See
https://fossa.com/blog/open-source-software-licenses-101-agpl-license/
for in-depth explanation.
In short, this license ensures that if this software is used in a
project, including a web application, the whole project needs to be
licensed under AGPL (or compatible license).
This was my original intent with GPLv3, but I was not aware that it does
not cover SaaS cases and such.

From https://www.gnu.org/licenses/license-list.en.html
> This is a free software, copyleft license. Its terms effectively consist of the terms of GPLv3, with an additional paragraph in section 13 to allow users who interact with the licensed software over a network to receive the source for that program. We recommend that developers consider using the GNU AGPL for any software which will commonly be run over a network.